+++ src/aosm/azext_aosm/cli_handlers/onboarding_sns_handler.py
@@ -9,28 +9,38 @@ from pathlib import Path
 from typing import Optional
 
 from knack.log import get_logger
+
 from azext_aosm.cli_handlers.onboarding_nfd_base_handler import OnboardingBaseCLIHandler
+from azext_aosm.common.command_context import CommandContext
 from azext_aosm.common.constants import (
     SNS_OUTPUT_FOLDER_FILENAME,
-    SNS_INPUT_FILENAME
+    SNS_INPUT_FILENAME,
+    SNS_DEFINITION_FOLDER_NAME,
+    SNS_TEMPLATE_FOLDER_NAME,
+    SNS_DEFINITION_TEMPLATE_FILENAME,
 )
-from azext_aosm.configuration_models.common_parameters_config import (
-    SNSCommonParametersConfig,
+from azext_aosm.common.utils import render_bicep_contents_from_j2, get_template_path
+from azext_aosm.configuration_models.onboarding_sns_input_config import (
+    OnboardingSNSInputConfig,
 )
-from azext_aosm.definition_folder.builder.bicep_builder import (
-    BicepDefinitionElementBuilder,
-)
+from azext_aosm.configuration_models.sns_parameters_config import SNSCommonParametersConfig
+from azext_aosm.definition_folder.builder.bicep_builder import BicepDefinitionElementBuilder
+from azext_aosm.definition_folder.builder.sns_deploy_input_builder import SNSDeploymentInputDefinitionElementBuilder
 from azext_aosm.definition_folder.builder.json_builder import (
     JSONDefinitionElementBuilder,
 )
-from azext_aosm.configuration_models.onboarding_sns_input_config import SiteNetworkServicePropertiesConfig
-logger = get_logger(__name__)
+from azext_aosm.definition_folder.reader.definition_folder import DefinitionFolder
+from azext_aosm.vendored_sdks import HybridNetworkManagementClient
+from azext_aosm.vendored_sdks.models import (
+    NetworkServiceDesignVersion,
+)
 
 
-class SNSCLIHandler(OnboardingBaseCLIHandler):
-    """CLI handler for SNS."""
+class OnboardingSNSCLIHandler(OnboardingBaseCLIHandler):
+    """CLI handler for deploying SNSs."""
 
-    config: SiteNetworkServicePropertiesConfig
+    config: OnboardingSNSInputConfig | SNSCommonParametersConfig
+    logger = get_logger(__name__)
 
     @property
     def default_config_file_name(self) -> str:
@@ -44,14 +54,19 @@ class SNSCLIHandler(OnboardingBaseCLIHandler):
 
     def build(self):
         """Build the definition."""
+        self.definition_folder_builder.add_element(self.build_resource_bicep())
+        self.definition_folder_builder.add_element(self.build_deploy_input())
         self.definition_folder_builder.add_element(self.build_all_parameters_json())
         self.definition_folder_builder.write()
 
-    def _get_input_config(self, input_config: Optional[dict] = None) -> SiteNetworkServicePropertiesConfig:
+    def _get_processor_list(self) -> list:
+        return []
+
+    def _get_input_config(self, input_config: Optional[dict] = None) -> OnboardingSNSInputConfig:
         """Get the configuration for the command."""
         if input_config is None:
             input_config = {}
-        return SiteNetworkServicePropertiesConfig(**input_config)
+        return OnboardingSNSInputConfig(**input_config)
 
     def _get_params_config(self, config_file: Path) -> SNSCommonParametersConfig:
         """Get the configuration for the command."""
@@ -61,8 +76,54 @@ class SNSCLIHandler(OnboardingBaseCLIHandler):
             params_dict = {}
         return SNSCommonParametersConfig(**params_dict)
 
-    def build_base_bicep(self) -> BicepDefinitionElementBuilder:
-        pass
+    def build_deploy_input(self) -> SNSDeploymentInputDefinitionElementBuilder:
+        """Pre-validate the build."""
+        nsdv = self._get_nsdv()
+        deployment_input_file = SNSDeploymentInputDefinitionElementBuilder(
+            Path(SNS_OUTPUT_FOLDER_FILENAME), nsdv.properties.nfvis_from_site
+        )
+        return deployment_input_file
+
+    def _get_nsdv(self) -> NetworkServiceDesignVersion:
+        """Get the existing NSDV resource object."""
+        self.logger.debug(
+            "Reading existing NSDV resource object %s from group %s",
+            self.config.nsd_reference.nsd_version,
+            self.config.nsd_reference.nsd_name
+        )
+        assert isinstance(self.aosm_client, HybridNetworkManagementClient)
+        nsdv_object = self.aosm_client.network_service_design_versions.get(
+            resource_group_name=self.config.nsd_reference.publisher_resource_group_name,
+            publisher_name=self.config.nsd_reference.publisher_name,
+            network_service_design_group_name=self.config.nsd_reference.nsd_name,
+            network_service_design_version_name=self.config.nsd_reference.nsd_version,
+        )
+        return nsdv_object
+
+    def build_resource_bicep(self) -> BicepDefinitionElementBuilder:
+        """Build the resource bicep file."""
+        template_path = get_template_path(
+            SNS_TEMPLATE_FOLDER_NAME, SNS_DEFINITION_TEMPLATE_FILENAME
+        )
+        params = {}
+
+        bicep_contents = render_bicep_contents_from_j2(
+            template_path, params
+        )
+        # Generate the nsd bicep file
+        bicep_file = BicepDefinitionElementBuilder(
+            Path(SNS_OUTPUT_FOLDER_FILENAME, SNS_DEFINITION_FOLDER_NAME), bicep_contents
+        )
+
+        return bicep_file
+
+    def deploy(self, command_context: CommandContext):
+        """Publish the definition contained in the specified definition folder."""
+        definition_folder = DefinitionFolder(
+            command_context.cli_options["definition_folder"]
+        )
+        assert isinstance(self.config, SNSCommonParametersConfig)
+        definition_folder.deploy(config=self.config, command_context=command_context)
 
     def build_all_parameters_json(self) -> JSONDefinitionElementBuilder:
         """Build all parameters json."""
@@ -76,18 +137,14 @@ class SNSCLIHandler(OnboardingBaseCLIHandler):
         )
         return base_file
 
-    def _get_processor_list(self):
-        # Provide an implementation for this method
-        pass
+    def build_base_bicep(self):
+        # TODO: Implement
+        raise NotImplementedError
 
     def build_artifact_list(self):
-        # Provide an implementation for this method
-        pass
+        # TODO: Implement this method
+        raise NotImplementedError
 
-    def build_manifest_bicep(self):
-        # Provide an implementation for this method
-        pass
-
-    def build_resource_bicep(self) -> BicepDefinitionElementBuilder:
-        # Provide an implementation for this method
-        pass
+    def build_manifest_bicep(self) -> BicepDefinitionElementBuilder:
+        # TODO: Implement this method
+        raise NotImplementedError