+++ src/aosm/HISTORY.rst
@@ -40,6 +40,9 @@ Release History
 * Fixed helm chart image parsing in the `az aosm nfd build` command for the `cnf` definition type. This means that the images can now be extracted correctly from the helm chart.
 * Fixed: infinite loop bug when retrying failed artifact uploads to the ACR
 * Added a check to determine if all base resources are already deployed. If they all are, the deployment of base bicep is skipped.
+* Added: Build and Deploy operations for site resource.
+    * `az aosm site build` to prepare files for a site.
+    * `az aosm site deploy` to deploy a site.
 
 1.0.0b4
 ++++++++